# openapi.yaml
openapi: 3.0.3
info:
  title: StitchAdmin 2.0 API
  description: |
    Modernisierte ERP-Lösung für Stickerei-Betriebe

    ## Authentifizierung
    Die meisten Endpunkte erfordern eine Session-basierte Authentifizierung über Flask-Login.
    Melden Sie sich zuerst über `/auth/login` an, um eine Session zu erhalten.

    ## Berechtigung
    Viele Endpunkte erfordern Admin-Rechte. Diese sind mit dem Tag `admin` gekennzeichnet.

    ## Datenformate
    - Alle Datumsangaben sind im ISO 8601 Format (YYYY-MM-DD oder YYYY-MM-DDTHH:MM:SS)
    - Preise sind als Float-Werte in Euro
    - IDs sind alphanumerische Strings oder Integers je nach Entität

    Erstellt von Hans Hahn - Alle Rechte vorbehalten
  version: 2.0.0
  contact:
    name: StitchAdmin Support
    email: admin@stitchadmin.local
servers:
  - url: http://localhost:5000
    description: Entwicklungsserver
  - url: https://stitchadmin.example.com
    description: Produktionsserver
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: session
      description: Session-Cookie nach Login
...
security:
  - sessionAuth: []
paths:
...
  /rechnung:
    get:
      tags:
        - Rechnungen
      summary: Rechnungs-Übersicht
      security: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: customer
          in: query
          schema:
            type: string
        - name: period
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Rechnungs-Liste HTML
  /rechnung/neu:
    post:
      tags:
        - Rechnungen
      summary: Neue Rechnung erstellen
      security: []
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                customer_id:
                  type: string
                positions:
                  type: array
                  items:
                    type: object
      responses:
        '302':
          description: Redirect zur Rechnung
  /rechnung/{rechnung_id}/pdf:
    get:
      tags:
        - Rechnungen
      summary: Rechnung als PDF
      security: []
      parameters:
        - name: rechnung_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: PDF-Datei
          content:
            application/pdf:
              schema:
                type: string
                format: binary
  /rechnung/{rechnung_id}/download:
    get:
      tags:
        - Rechnungen
      summary: ZUGFeRD-PDF herunterladen
      security: []
      parameters:
        - name: rechnung_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: ZUGFeRD PDF-Datei
          content:
            application/pdf:
              schema:
                type: string
                format: binary

  # ==================== EINSTELLUNGEN ====================
...